# Remove `rising_factorial(::Int, ::Int)`