The version from AbstractAlgebra is actually faster and avoids allocations.

And this gets rid of one instances of type piracy, see #1495

Before:
```
julia> @benchmark sum(i->rising_factorial(10,i), 1:12)
BenchmarkTools.Trial: 10000 samples with 210 evaluations.
 Range (min … max):  352.976 ns … 519.090 μs  ┊ GC (min … max):  0.00% … 53.77%
 Time  (median):     359.324 ns               ┊ GC (median):     0.00%
 Time  (mean ± σ):   768.972 ns ±  14.411 μs  ┊ GC (mean ± σ):  28.19% ±  1.50%

     ▁     ▂█▃
  ▁▂▇█▅▂▁▂▄███▅▂▂▂▂▂▃▃▂▂▂▂▂▂▂▂▂▁▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  353 ns           Histogram: frequency by time          389 ns <

 Memory estimate: 192 bytes, allocs estimate: 12.
```

After:
```
julia> @benchmark sum(i->rising_factorial(10,i), 1:12)
BenchmarkTools.Trial: 10000 samples with 992 evaluations.
 Range (min … max):  38.348 ns … 144.111 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     38.516 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   39.126 ns ±   2.919 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █▄▁▁▂▃▁  ▁▁▂▁                                                ▁
  █████████████▅▅▄▅▁▅▆▆▅▄▅▅▅▅▅▄▄▄▄▅▄▅▃▄▃▄▅▅▄▄▅▅▅▅▄▅▄▄▅▅▅▅▅▄▅▅▅ █
  38.3 ns       Histogram: log(frequency) by time      54.1 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```